refactor(xsync): move all sync utilities to `xsync`
