+++ internal/xsync/xsync.go
@@ -1,0 +1,2 @@
+// Package xsync contains synchronization utilities.
+package xsync
